pyproject.toml
[project]
name = "resilient-redis-queue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "redis>=5.1.1",
    "requests>=2.32.3",
]
